refactor(proto): move proto api to v1
