+++ crates/jstz_proto/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod api;
 mod error;
 
 pub mod context;
@@ -8,3 +7,5 @@ pub mod operation;
 pub mod receipt;
 pub mod request_logger;
 pub use error::{Error, Result};
+
+pub mod runtime;
